requirements.txt
numpy
scipy
matplotlib
tqdm
scikit-learn
scikit-image
pillow
SimpleITK
tensorflow-gpu
keras==2.1.6
h5py
